Fix for variable scope issue
The v2.8.6 method definition for _autoclearCookies has an issue.

The value for domains is only ever set once at the top of the method, and is overwritten by any individual row’s domain property in the cookie table blocks.

This patch ensures that any override is scoped to that particular row only.